Began UI Formatting and CSS Stylesheet
The css stylesheet has been created and is stored under a new folder named static. The CSS sheet mostly contains classes related to the palette being used for the UI and some formatting for input boxes. Some html files have been changed to allow for classes to be added to specific fields. Most notable change is using the render_field() for each item in a form as opposed to generating an entire form using render_form()