applying now works
Can now apply from a job listing, most of the import from the user profile is working, need to get the select fields default to work.
Now shows which fields are required on forms.
Did not test whether the binary file upload to the database is retrievable and haven't made any methods to make retrievable yet.