Apply button redirect to login
Fixed the issue with the apply button not redirecting to the login page if the user is not logged in. If the user isn't logged in it creates a button with an href to the login page. If the user IS logged in (authenticated) the button is created as it was originally. Basically just an if else case